// frontend/src/app/api/model/createCourseDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateCourseDto { 
    courseCode?: string | null;
    title?: string | null;
    description?: string | null;
    award?: string | null;
    durationSemesters?: number | null;
    isActive?: boolean;
}